refactoring blockchain package
