+++ src/blockchain/block.test.js
@@ -1,4 +1,4 @@
-import Block from './Block.js';
+import Block from './block';
 
 describe('Block', () => {
     let timestamp;
