+++ src/blockchain/index.js
@@ -1,0 +1,5 @@
+import Block from './block';
+import Blockchain from './blockchain';
+
+export { Block };
+export default Blockchain;
